4. _realloc

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -20,7 +20,7 @@
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 {
 	void *new_ptr;
-	unsigned int i, copy_size;
+	size_t i, copy_size;
 
 	if (new_size == 0)
 	{
@@ -32,12 +32,12 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 	{
 		new_ptr = malloc(new_size);
 		if (new_ptr == NULL)
-		return (NULL);
+			return (NULL);
 		return (new_ptr);
 	}
 
 	if (new_size == old_size)
-	return (ptr);
+		return (ptr);
 
 	new_ptr = malloc(new_size);
 
@@ -47,7 +47,7 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 	copy_size = old_size < new_size ? old_size : new_size;
 
 	for (i = 0; i < copy_size; i++)
-	((char *)new_ptr)[i] = ((char *)ptr)[i];
+		((char *)new_ptr)[i] = ((char *)ptr)[i];
 
 	free(ptr);
